Changed: Allow Motion blur in Freeze Frame mode. Changed: More consistency changes and code cleanup in Motion Blur. Fixed: Motion Blur on alpha masked objects. Fixed: Motion Blur on alpha masked rigged meshes. Fixed: Motion Blur breaking on fitted mesh, fixed once and for all.
--HG--
branch : Black Dragon

+++ indra/newview/app_settings/shaders/class1/deferred/motionBlurF.glsl
@@ -35,7 +35,7 @@ uniform sampler2DRect diffuseRect;
 uniform sampler2DRect normalMap;
 
 uniform float time_step;
-uniform float mblur_strength;
+uniform int mblur_strength;
 
 VARYING vec4 vary_fragcoord;
 uniform vec2 screen_res;
